check to see if already downloaded

+++ commentFetch.py
@@ -18,53 +18,51 @@ class InfoFetch():
                                    username=redditDict['username'])
 
     def getSubmission(self, inSubmission, inLimit=100):
-        
-        retList = []
-        commentDict = {}
+        if(not os.path.exists('{0}.json'.format(inSubmission))):
+            retList = []
+            commentDict = {}
 
-        subreddit = self.reddit.subreddit(inSubmission)
-        subList = [submission.id for submission in subreddit.top(limit=inLimit)] #make a list of all the subreddit ids
-        
-        print(subList)
+            subreddit = self.reddit.subreddit(inSubmission)
+            subList = [submission.id for submission in subreddit.top(limit=inLimit)] #make a list of all the subreddit ids
+            
+            print(subList)
 
-        for i in subList:
-            tempStr = 'https://api.pushshift.io/reddit/submission/comment_ids/' + i
-            r = requests.get(tempStr)
-            j = r.json() 
-            commentIds = j["data"]
-            commentDict[i] = commentIds #make a dict with submission as key, and comment ids list as value
-            print('retrived', i)
-            time.sleep(.3)
+            for i in subList:
+                tempStr = 'https://api.pushshift.io/reddit/submission/comment_ids/' + i
+                r = requests.get(tempStr)
+                j = r.json() 
+                commentIds = j["data"]
+                commentDict[i] = commentIds #make a dict with submission as key, and comment ids list as value
+                print('retrived', i)
+                time.sleep(.3)
 
-        trueDict = {}
-        for key, val in commentDict.items():
-            if len(val) > 1000:
-                tempList = []
-                loops = len(val) // 1000
-                for i in range(loops):
-                    tempStr = 'https://api.pushshift.io/reddit/comment/search?ids=' + ','.join(val[1000*i:1000*(i+1)])
+            trueDict = {}
+            for key, val in commentDict.items():
+                if len(val) > 1000:
+                    tempList = []
+                    loops = len(val) // 1000
+                    for i in range(loops):
+                        tempStr = 'https://api.pushshift.io/reddit/comment/search?ids=' + ','.join(val[1000*i:1000*(i+1)])
+                        r = requests.get(tempStr)
+                        j = r.json()
+                        tempList += [i['body'] for i in j['data']]
+                        time.sleep(.3)
+                    tempStr = 'https://api.pushshift.io/reddit/comment/search?ids=' + ','.join(val[1000*(i+1):len(val)])
                     r = requests.get(tempStr)
                     j = r.json()
                     tempList += [i['body'] for i in j['data']]
+                    trueDict[key] = tempList
                     time.sleep(.3)
-                tempStr = 'https://api.pushshift.io/reddit/comment/search?ids=' + ','.join(val[1000*(i+1):len(val)])
-                r = requests.get(tempStr)
-                j = r.json()
-                tempList += [i['body'] for i in j['data']]
-                trueDict[key] = tempList
-                time.sleep(.3)
-            else:
-                tempStr = 'https://api.pushshift.io/reddit/comment/search?ids=' + ','.join(val)
-                r = requests.get(tempStr)
-                j = r.json()
-                trueDict[key] = [i['body'] for i in j['data']] #make a dictionary with submisison as id, and comment list as value
-                time.sleep(.3)
-            print('got comments', key)
+                else:
+                    tempStr = 'https://api.pushshift.io/reddit/comment/search?ids=' + ','.join(val)
+                    r = requests.get(tempStr)
+                    j = r.json()
+                    trueDict[key] = [i['body'] for i in j['data']] #make a dictionary with submisison as id, and comment list as value
+                    time.sleep(.3)
+                print('got comments', key)
 
-        with open('{0}.json'.format(inSubmission), 'w') as newFile:
-            json.dump(trueDict, newFile, indent=4, separators=(',', ': '), ensure_ascii=False)
-
-        return retList
+            with open('{0}.json'.format(inSubmission), 'w') as newFile:
+                json.dump(trueDict, newFile, indent=4, separators=(',', ': '), ensure_ascii=False)
 
 
 if __name__ == '__main__':
